improved test structure
