+++ tests/cfdp/pdus/test_pdu_wrapper.py
@@ -1,0 +1,6 @@
+from unittest import TestCase
+
+
+class TestPduWrapper(TestCase):
+    def test_wrapper(self):
+        pass
